Refactor : Facade 패턴으로 바꾸게 되어 LikeAccessService 삭제 및 관련 test 코드 수정
