--- src/main/java/org/guzzing/studayserver/domain/like/service/LikeAccessService.java
@@ -1,6 +1,0 @@
-package org.guzzing.studayserver.domain.like.service;
-
-public interface LikeAccessService {
-
-    boolean isLiked(Long academyId, Long memberId);
-}
